Moving diagnostics code to diagnostics package
